Shuffle content to its own spot
