+++ src/content/Welcome.js
@@ -5,10 +5,10 @@ import Map from '../components/Map';
 
 export const Component = () => {
   return (
-    <Article title='Welcome' content={import ('../assets/markdown/Welcome.md') }>
+    <Article title='Welcome' content={import ('./Welcome.md') }>
       <Map
         name="Hinterlands of Fate's End"
-        image={import('../assets/maps/fatesEnd.jpg')}
+        image={import('./fatesEnd.jpg')}
         initialScale={0.5}
         minScale={0.1}
         maxScale={3}
